Move MockedAutoResolver class to a separate file
