+++ test/Tethos.Tests/SUT/MockedAutoResolver.cs
@@ -1,0 +1,22 @@
+﻿namespace Tethos.Tests.SUT
+{
+    using System;
+    using Castle.MicroKernel;
+
+    internal class MockedAutoResolver : BaseAutoResolver
+    {
+        public MockedAutoResolver(IKernel kernel)
+            : base(kernel)
+        {
+        }
+
+        public override object MapToMock(Type targetType, object targetObject, Arguments constructorArguments) =>
+            new MapToMockArguments
+            {
+                TargetType = targetType,
+                TargetObject = targetObject,
+                ConstructorArguments = constructorArguments,
+            };
+    }
+
+}
